Revert "build: Use po4a --srcdir and --builddir"
This reverts commit 822ea85c7c16d45f932f9c86f5ac0a04a8c304c2.

The squeeze version of po4a (0.40.2-1) does not work as
expected with --srcdir. The switch to srcdir happens after
the check of the existence of the master file and it thus
fails with “The 'deb.5' master file does not exist.”. The
wheezy version (0.41-1) is not affected by this problem.

Conflicts:

	man/Makefile.am
	man/po/po4a.cfg